Camera and Gallery
The camera function and gallery function work. However, I haven't
figured out how to use those picture or display the pictures
picked/taken, but it's a start. Also, I figured out that when the camera
function is called it actually saves the picture in the users gallery.
It could be good and bad news, I haven't made my mind up which one it
is. lol But anyways, yeah let me know what you think.

P.S. There is a lot of random code in here that I have gotten from the
internet and tried to convert it to make our code run but a lot of it
doesn't really work like it should, but whatever